+++ opentelemetry-otlp/src/exporter/mod.rs
@@ -18,11 +18,30 @@ use std::time::Duration;
 /// <https://github.com/open-telemetry/opentelemetry-specification/blob/main/specification/protocol/exporter.md#endpoint-urls-for-otlphttp>
 pub const OTEL_EXPORTER_OTLP_ENDPOINT: &str = "OTEL_EXPORTER_OTLP_ENDPOINT";
 /// Default target to which the exporter is going to send signals.
-pub const OTEL_EXPORTER_OTLP_ENDPOINT_DEFAULT: &str = "https://localhost:4317";
+pub const OTEL_EXPORTER_OTLP_ENDPOINT_DEFAULT: &str = OTEL_EXPORTER_OTLP_HTTP_ENDPOINT_DEFAULT;
+/// Protocol to use when sending
+pub const OTEL_EXPORTER_OTLP_PROTOCOL: &str = "OTEL_EXPORTER_OTLP_PROTOCOL";
+
+/// Default protocol, if http-proto is enabled then use it, otherwise use grpc
+#[cfg(feature = "http-proto")]
+pub const OTEL_EXPORTER_OTLP_PROTOCOL_DEFAULT: &str = OTEL_EXPORTER_OTLP_PROTOCOL_HTTP_PROTOBUF;
+#[cfg(all(
+    any(feature = "grpc-tonic", feature = "grpcio"),
+    not(feature = "http-proto")
+))]
+pub const OTEL_EXPORTER_OTLP_PROTOCOL_DEFAULT: &str = OTEL_EXPORTER_OTLP_PROTOCOL_GRPC;
+
+const OTEL_EXPORTER_OTLP_PROTOCOL_HTTP_PROTOBUF: &str = "http/protobuf";
+const OTEL_EXPORTER_OTLP_PROTOCOL_GRPC: &str = "grpc";
+
 /// Max waiting time for the backend to process each signal batch, defaults to 10 seconds.
 pub const OTEL_EXPORTER_OTLP_TIMEOUT: &str = "OTEL_EXPORTER_OTLP_TIMEOUT";
 /// Default max waiting time for the backend to process each signal batch.
 pub const OTEL_EXPORTER_OTLP_TIMEOUT_DEFAULT: u64 = 10;
+
+// Endpoints per protocol https://github.com/open-telemetry/opentelemetry-specification/blob/main/specification/protocol/exporter.md
+const OTEL_EXPORTER_OTLP_GRPC_ENDPOINT_DEFAULT: &str = "http://localhost:4317";
+const OTEL_EXPORTER_OTLP_HTTP_ENDPOINT_DEFAULT: &str = "http://localhost:4318";
 
 #[cfg(feature = "grpc-sys")]
 pub(crate) mod grpcio;
@@ -46,11 +65,30 @@ pub struct ExportConfig {
 
 impl Default for ExportConfig {
     fn default() -> Self {
+        let protocol = default_protocol();
+
         ExportConfig {
-            endpoint: OTEL_EXPORTER_OTLP_ENDPOINT_DEFAULT.to_string(),
-            protocol: Protocol::Grpc,
+            endpoint: default_endpoint(protocol),
+            protocol,
             timeout: Duration::from_secs(OTEL_EXPORTER_OTLP_TIMEOUT_DEFAULT),
         }
+    }
+}
+
+/// default protocol based on enabled features
+fn default_protocol() -> Protocol {
+    match OTEL_EXPORTER_OTLP_PROTOCOL_DEFAULT {
+        OTEL_EXPORTER_OTLP_PROTOCOL_HTTP_PROTOBUF => Protocol::HttpBinary,
+        OTEL_EXPORTER_OTLP_PROTOCOL_GRPC => Protocol::Grpc,
+        _ => Protocol::HttpBinary,
+    }
+}
+
+/// default endpoint for protocol
+fn default_endpoint(protocol: Protocol) -> String {
+    match protocol {
+        Protocol::Grpc => OTEL_EXPORTER_OTLP_GRPC_ENDPOINT_DEFAULT.to_string(),
+        Protocol::HttpBinary => OTEL_EXPORTER_OTLP_HTTP_ENDPOINT_DEFAULT.to_string(),
     }
 }
 
@@ -137,9 +175,20 @@ impl<B: HasExportConfig> WithExportConfig for B {
     }
 
     fn with_env(mut self) -> Self {
+        let protocol = match std::env::var(OTEL_EXPORTER_OTLP_PROTOCOL)
+            .unwrap_or(OTEL_EXPORTER_OTLP_PROTOCOL_DEFAULT.to_string())
+            .as_str()
+        {
+            OTEL_EXPORTER_OTLP_PROTOCOL_HTTP_PROTOBUF => Protocol::HttpBinary,
+            OTEL_EXPORTER_OTLP_PROTOCOL_GRPC => Protocol::Grpc,
+            _ => default_protocol(),
+        };
+
+        self.export_config().protocol = protocol;
+
         let endpoint = match std::env::var(OTEL_EXPORTER_OTLP_ENDPOINT) {
             Ok(val) => val,
-            Err(_) => OTEL_EXPORTER_OTLP_ENDPOINT_DEFAULT.to_string(),
+            Err(_) => default_endpoint(protocol),
         };
         self.export_config().endpoint = endpoint;
 
@@ -162,38 +211,134 @@ impl<B: HasExportConfig> WithExportConfig for B {
 #[cfg(test)]
 #[cfg(feature = "grpc-tonic")]
 mod tests {
+    const LOCK_POISONED_MESSAGE: &'static str =
+        "one of the other pipeline builder from env tests failed";
     use crate::exporter::{
-        WithExportConfig, OTEL_EXPORTER_OTLP_ENDPOINT, OTEL_EXPORTER_OTLP_TIMEOUT,
-        OTEL_EXPORTER_OTLP_TIMEOUT_DEFAULT,
+        default_endpoint, default_protocol, WithExportConfig, OTEL_EXPORTER_OTLP_ENDPOINT,
+        OTEL_EXPORTER_OTLP_GRPC_ENDPOINT_DEFAULT, OTEL_EXPORTER_OTLP_HTTP_ENDPOINT_DEFAULT,
+        OTEL_EXPORTER_OTLP_PROTOCOL_GRPC, OTEL_EXPORTER_OTLP_PROTOCOL_HTTP_PROTOBUF,
+        OTEL_EXPORTER_OTLP_TIMEOUT, OTEL_EXPORTER_OTLP_TIMEOUT_DEFAULT,
     };
-    use crate::new_exporter;
+    use crate::{new_exporter, Protocol, OTEL_EXPORTER_OTLP_PROTOCOL};
+    use std::sync::Mutex;
+
+    // Make sure env tests are not running concurrently
+    static ENV_LOCK: Mutex<usize> = Mutex::new(0);
 
     #[test]
     fn test_pipeline_builder_from_env_default_vars() {
-        let expected_endpoint = "https://otlp_endpoint:4317";
-        std::env::set_var(OTEL_EXPORTER_OTLP_ENDPOINT, expected_endpoint);
-        std::env::set_var(OTEL_EXPORTER_OTLP_TIMEOUT, "bad_timeout");
-
-        let mut exporter_builder = new_exporter().tonic().with_env();
-        assert_eq!(exporter_builder.exporter_config.endpoint, expected_endpoint);
-
-        exporter_builder = new_exporter().tonic().with_env();
+        let _env_lock = ENV_LOCK.lock().expect(LOCK_POISONED_MESSAGE);
+        let exporter_builder = new_exporter().tonic().with_env();
+        assert_eq!(
+            exporter_builder.exporter_config.protocol,
+            default_protocol()
+        );
+        assert_eq!(
+            exporter_builder.exporter_config.endpoint,
+            default_endpoint(default_protocol())
+        );
         assert_eq!(
             exporter_builder.exporter_config.timeout,
             std::time::Duration::from_secs(OTEL_EXPORTER_OTLP_TIMEOUT_DEFAULT)
         );
+    }
 
+    #[test]
+    fn test_pipeline_builder_from_env_endpoint() {
+        let _env_lock = ENV_LOCK.lock().expect(LOCK_POISONED_MESSAGE);
+        std::env::set_var(OTEL_EXPORTER_OTLP_ENDPOINT, "http://example.com");
+        let exporter_builder = new_exporter().tonic().with_env();
+        assert_eq!(
+            exporter_builder.exporter_config.endpoint,
+            "http://example.com"
+        );
+        std::env::remove_var(OTEL_EXPORTER_OTLP_ENDPOINT);
+        assert!(std::env::var(OTEL_EXPORTER_OTLP_ENDPOINT).is_err());
+    }
+
+    #[test]
+    fn test_pipeline_builder_from_env_protocol_http_protobuf() {
+        let _env_lock = ENV_LOCK.lock().expect(LOCK_POISONED_MESSAGE);
+        std::env::set_var(
+            OTEL_EXPORTER_OTLP_PROTOCOL,
+            OTEL_EXPORTER_OTLP_PROTOCOL_HTTP_PROTOBUF,
+        );
+        let exporter_builder = new_exporter().tonic().with_env();
+        assert_eq!(
+            exporter_builder.exporter_config.protocol,
+            Protocol::HttpBinary
+        );
+        assert_eq!(
+            exporter_builder.exporter_config.endpoint,
+            OTEL_EXPORTER_OTLP_HTTP_ENDPOINT_DEFAULT
+        );
+
+        std::env::remove_var(OTEL_EXPORTER_OTLP_PROTOCOL);
+        assert!(std::env::var(OTEL_EXPORTER_OTLP_PROTOCOL).is_err());
+    }
+
+    #[test]
+    fn test_pipeline_builder_from_env_protocol_grpc() {
+        let _env_lock = ENV_LOCK.lock().expect(LOCK_POISONED_MESSAGE);
+        std::env::set_var(
+            OTEL_EXPORTER_OTLP_PROTOCOL,
+            OTEL_EXPORTER_OTLP_PROTOCOL_GRPC,
+        );
+        let exporter_builder = new_exporter().tonic().with_env();
+        assert_eq!(exporter_builder.exporter_config.protocol, Protocol::Grpc);
+        assert_eq!(
+            exporter_builder.exporter_config.endpoint,
+            OTEL_EXPORTER_OTLP_GRPC_ENDPOINT_DEFAULT
+        );
+
+        std::env::remove_var(OTEL_EXPORTER_OTLP_PROTOCOL);
+        assert!(std::env::var(OTEL_EXPORTER_OTLP_PROTOCOL).is_err());
+    }
+
+    #[test]
+    fn test_pipeline_builder_from_env_bad_protocol() {
+        let _env_lock = ENV_LOCK.lock().expect(LOCK_POISONED_MESSAGE);
+        std::env::set_var(OTEL_EXPORTER_OTLP_PROTOCOL, "bad_protocol");
+        let exporter_builder = new_exporter().tonic().with_env();
+        assert_eq!(
+            exporter_builder.exporter_config.protocol,
+            default_protocol()
+        );
+        assert_eq!(
+            exporter_builder.exporter_config.endpoint,
+            default_endpoint(default_protocol())
+        );
+
+        std::env::remove_var(OTEL_EXPORTER_OTLP_PROTOCOL);
+        assert!(std::env::var(OTEL_EXPORTER_OTLP_PROTOCOL).is_err());
+    }
+
+    #[test]
+    fn test_pipeline_builder_from_env_timeout() {
+        let _env_lock = ENV_LOCK.lock().expect(LOCK_POISONED_MESSAGE);
         std::env::set_var(OTEL_EXPORTER_OTLP_TIMEOUT, "60");
-
-        exporter_builder = new_exporter().tonic().with_env();
+        let exporter_builder = new_exporter().tonic().with_env();
         assert_eq!(
             exporter_builder.exporter_config.timeout,
             std::time::Duration::from_secs(60)
         );
 
-        std::env::remove_var(OTEL_EXPORTER_OTLP_ENDPOINT);
         std::env::remove_var(OTEL_EXPORTER_OTLP_TIMEOUT);
-        assert!(std::env::var(OTEL_EXPORTER_OTLP_ENDPOINT).is_err());
+        assert!(std::env::var(OTEL_EXPORTER_OTLP_TIMEOUT).is_err());
+    }
+
+    #[test]
+    fn test_pipeline_builder_from_env_bad_timeout() {
+        let _env_lock = ENV_LOCK.lock().expect(LOCK_POISONED_MESSAGE);
+        std::env::set_var(OTEL_EXPORTER_OTLP_TIMEOUT, "bad_timeout");
+
+        let exporter_builder = new_exporter().tonic().with_env();
+        assert_eq!(
+            exporter_builder.exporter_config.timeout,
+            std::time::Duration::from_secs(OTEL_EXPORTER_OTLP_TIMEOUT_DEFAULT)
+        );
+
+        std::env::remove_var(OTEL_EXPORTER_OTLP_TIMEOUT);
         assert!(std::env::var(OTEL_EXPORTER_OTLP_TIMEOUT).is_err());
     }
 }